% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/code.R
\name{beak}
\alias{beak}
\title{Parse a darwin core file}
\usage{
beak(file)
}
\arguments{
\item{file}{A darwin core file}
}
\description{
Parse a darwin core file
}
\examples{
\dontrun{
file <- system.file("examples", "example_simple.xml", package = "finch")
beak(file)
file <- system.file("examples", "example_simple_fossil.xml", package = "finch")
beak(file)
}
}

