% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{finch-defunct}
\alias{finch-defunct}
\title{Defunct functions in finch}
\description{
\itemize{
\item \code{dwca_cache_delete}: Defunt - see \link{dwca_cache}
\item \code{dwca_cache_delete_all}: Defunt - see \link{dwca_cache}
\item \code{dwca_cache_details}: Defunt - see \link{dwca_cache}
\item \code{dwca_cache_list}: Defunt - see \link{dwca_cache}
}
}
