% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gbif_validate.R
\name{gbif_validate}
\alias{gbif_validate}
\title{Validate a Darwin Core Archive using GBIF's validator}
\usage{
gbif_validate(object)
}
\arguments{
\item{object}{Either a url or a file}
}
\description{
Validate a Darwin Core Archive using GBIF's validator
}
\examples{
\dontrun{
gbif_validate()
}
}

