% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_validate.R
\name{schema_validate}
\alias{schema_validate}
\title{Validate a Darwin Core Archive file using the DCA schema}
\usage{
schema_validate(file_path, schema_path)
}
\arguments{
\item{file}{(character) A file path. Must be a Darwin Core Archive file}
}
\description{
Validate a Darwin Core Archive file using the DCA schema
}
\examples{
\dontrun{
file <- system.file("examples", "example_simple.xml", package = "finch")
file <- system.file("examples", "example_simple_fossil.xml",
  package = "finch")
file <- system.file("examples", "example_classes_observation.xml",
  package = "finch")
schema_validate(file)

x <- system.file("examples", "example_bad1.xml", package = "finch")
# schema_validate(x)

schema_path <- system.file("schemas/tdwg_dwcterms.xsd", package = "finch")

# simple dwc file
file_path <- system.file("examples/example_simple.xml", package = "finch")
schema_path <- system.file("schemas/tdwg_dwc_simple.xsd", package = "finch")
schema_validate(file, schema_path)

# classes dwc file
file_path <- system.file("examples", "example_classes_specimen.xml", package = "finch")
schema_path <- system.file("schemas/tdwg_dwc_class_terms.xsd", package = "finch")
schema_validate(file, schema_path)
}
}
\keyword{internal}

