% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{Parse a simple darwin core file}
\usage{
simple(file)
}
\arguments{
\item{file}{A simple darwin core file in XML format}
}
\description{
Parse a simple darwin core file
}
\examples{
\dontrun{
file <- system.file("examples", "example_simple.xml", package = "finch")
simple(file)
file <- system.file("examples", "example_simple_fossil.xml", package = "finch")
simple(file)
file <- system.file("examples", "example_simple_fossil.xml", package = "finch")
simple(file)
}
}

