% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple_read}
\alias{simple_read}
\title{Parse a simple darwin core file}
\usage{
simple_read(file)
}
\arguments{
\item{file}{(character) A path to a single simple Darwin Core file in XML format}
}
\description{
Parse a simple darwin core file
}
\examples{
\dontrun{
file <- system.file("examples", "example_simple.xml", package = "finch")
simple_read(file)
file <- system.file("examples", "example_simple_fossil.xml", package = "finch")
simple_read(file)
file <- system.file("examples", "example_simple_fossil.xml", package = "finch")
simple_read(file)
}
}

