% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.location.R
\name{as.location}
\alias{as.location}
\alias{as.location.character}
\alias{as.location.location}
\alias{print.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, ...)

\method{as.location}{character}(x, ...)

\method{as.location}{location}(x, ...)

\method{print}{location}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}
\examples{
\donttest{
# A zip file
file <- "~/Downloads/0000154-150116162929234.zip"
as.location(file)

# A directory
dir <- "~/Downloads/0000154-150116162929234/"
as.location(dir)

# A URL
url <- "http://ecat-dev.gbif.org/repository/vernaculars/vernacular_registry_dwca_3.zip"
as.location(url)
}
}

