% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.location.R
\name{as.location}
\alias{as.location}
\alias{as.location.character}
\alias{as.location.location}
\alias{print.location}
\title{Convert a path or URL to a location object}
\usage{
as.location(x, ...)

\method{as.location}{character}(x, ...)

\method{as.location}{location}(x, ...)

\method{print}{location}(x, ...)
}
\arguments{
\item{x}{Input, a path or URL}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object
}
\examples{
# A zip file
file <- system.file("examples/0000154-150116162929234.zip",
  package = "finch")
as.location(file)

# A directory
dir <- system.file("examples/0000154-150116162929234",
  package = "finch")
as.location(dir)

# A URL
as.location("https://httpbin.org/get")
}
