% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dwca.R
\name{dwca}
\alias{dwca}
\title{Parse Darwin Core Archive}
\usage{
dwca(input, read = FALSE, ...)
}
\arguments{
\item{input}{(character) Path to local zip file, directory, or a url.}

\item{read}{(logical) Whether or not to read in data files. If \code{FALSE}, we
give back paths to files only. Default: \code{FALSE}}

\item{...}{Further args passed on to \code{\link[data.table]{fread}}}
}
\description{
Parse Darwin Core Archive
}
\details{
Note that sometimes file reads fail. We use \code{\link[data.table]{fread}}
internally, which is very fast, but can fail sometimes. If so, try reading in the
data manually.
}
\examples{
\dontrun{
dir <- system.file("examples", "0000154-150116162929234", package = "finch")

# Don't read data in
x <- dwca(dir, read=FALSE)
x
x$files
x$highmeta
x$dataset_meta[[1]]
x$data

# Read data
x <- dwca(dir, read=TRUE)
x
head( x$data[[1]] )

# Can pass in a zip file
zip <- system.file("examples", "0000154-150116162929234.zip", package = "finch")
out <- dwca(zip)
out$files
out$highmeta
out$emlmeta
out$dataset_meta

# Can pass in zip file as a url
url <-
"https://github.com/ropensci/finch/blob/master/inst/examples/0000154-150116162929234.zip?raw=true"
(out <- dwca(url))
}
}

