% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dwca.R
\name{dwca}
\alias{dwca}
\title{Parse Darwin Core Archive}
\usage{
dwca(input, read = FALSE, ...)
}
\arguments{
\item{input}{Path to local zip file, directory, or a url.}

\item{read}{(logical)}

\item{...}{Further args passed on to \code{\link[data.table]{fread}}}
}
\description{
Parse Darwin Core Archive
}
\details{
Note that sometimes file reads fail. We use \code{\link[data.table]{fread}}
internally, which is very fast, but can fail sometimes. If so, try reading in the
data manually.
}
\examples{
\dontrun{
dir <- "~/Downloads/0009933-141123120432318/"

# Don't read data in
x <- dwca(dir, read=FALSE)
x
x$files
x$highmeta
x$dataset_meta[[1]]
x$data

# Read data
x <- dwca(dir, read=TRUE)
x
head( x$data[[1]] )

out <- dwca("~/Downloads/0000143-150116162929234/")
out
dwca("~/Downloads/0000143-150116162929234/", TRUE)
dwca("~/Downloads/0000154-150116162929234/", TRUE)

# Can pass in a zip file
out <- dwca("~/Downloads/0000154-150116162929234.zip")

# Can pass in zip file as a url
x <- "http://ecat-dev.gbif.org/repository/vernaculars/vernacular_registry_dwca_3.zip"
out <- dwca(x)

# This one fails on file read
out <- dwca("~/Downloads/0000153-150116162929234/", TRUE)
}
}

