% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finch-package.R
\docType{package}
\name{finch-package}
\alias{finch-package}
\alias{finch}
\title{Parse Darwin Core Archive files}
\description{
Parse Darwin Core Archive files
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
